Add Stackoverflow-sourced solution to print u128
This divides the higher and lower parts of the integer
(High & low as divided in base 10.) First, the number
is divided to the highest power of 10 that is less than
the largest number in unsigned 64-bit integers. Then,
it prints the quotient of x divided by that power of 10
(high part.) Next, it prints the remainder (lower part).