Sort (#43)
* Try out with a freq vector

* Working, c++, sort

* Use a shared pointer to reduce copy overhead

In some instances, c++ implementation of sort
creates many copies of the Compare object.
To reduce making deep copies, it is desirable
to store pointers to the context-holding
structures. In C++, it is desirable to use
smart pointers rather than raw pointers.
Since many owners may exist during the sorting
shared ownership must be enforced. In C++,
`shared_ptr` is a canonical way to deal with
shared ownership.